rename users classes
