+++ app/Http/Controllers/Back/AdministratorsController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers\Back;
 
+use App\Models\Administrator;
 use App\Services\Auth\Back\Enums\UserRole;
 use App\Services\Auth\Back\Enums\UserStatus;
 use App\Services\Auth\Back\Events\UserCreated;
-use App\Services\Auth\Back\User;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -16,21 +16,21 @@ class AdministratorsController
 
     public function index()
     {
-        $users = User::all();
+        $users = Administrator::all();
 
         return view('back.administrators.index', compact('users'));
     }
 
     public function create()
     {
-        return view('back.administrators.create', ['user' => new User()]);
+        return view('back.administrators.create', ['user' => new Administrator()]);
     }
 
     public function store(Request $request)
     {
         $this->validate($request, $this->validationRules());
 
-        $user = new User();
+        $user = new Administrator();
 
         $user->email = $request->get('email');
         $user->first_name = $request->get('first_name');
@@ -56,7 +56,7 @@ class AdministratorsController
 
     public function edit($id)
     {
-        $user = User::findOrFail($id);
+        $user = Administrator::findOrFail($id);
 
         return view('back.administrators.edit', compact('user'));
     }
@@ -65,7 +65,7 @@ class AdministratorsController
     {
         $this->validate($request, $this->validationRules());
 
-        $user = User::findOrFail($id);
+        $user = Administrator::findOrFail($id);
 
         $user->email = $request->get('email');
         $user->first_name = $request->get('first_name');
@@ -86,7 +86,7 @@ class AdministratorsController
 
     public function activate($id)
     {
-        $user = User::findOrFail($id);
+        $user = Administrator::findOrFail($id);
 
         $user->activate();
 
@@ -99,7 +99,7 @@ class AdministratorsController
 
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
+        $user = Administrator::findOrFail($id);
 
         $eventDescription = $this->getEventDescriptionFor('deleted', $user);
 
@@ -111,7 +111,7 @@ class AdministratorsController
         return redirect()->action('Back\AdministratorsController@index');
     }
 
-    protected function getEventDescriptionFor(string $event, User $user): string
+    protected function getEventDescriptionFor(string $event, Administrator $user): string
     {
         $name = sprintf(
             '<a href="%s">%s</a>',
