+++ database/seeds/AdministratorSeeder.php
@@ -1,14 +1,14 @@
 <?php
 
+use App\Models\Administrator;
 use App\Services\Auth\Back\Enums\UserRole;
 use App\Services\Auth\Back\Enums\UserStatus;
-use App\Services\Auth\Back\User;
 
-class BackUserSeeder extends DatabaseSeeder
+class AdministratorSeeder extends DatabaseSeeder
 {
     public function run()
     {
-        $this->truncate((new User())->getTable());
+        $this->truncate((new Administrator())->getTable());
 
         collect([
             ['Alex', 'Vanderbist'],
@@ -21,7 +21,7 @@ class BackUserSeeder extends DatabaseSeeder
         ])->each(function ($name) {
             [$firstName, $lastName] = $name;
 
-            $this->createBackUser([
+            $this->createAdministrator([
                 'first_name' => $firstName,
                 'last_name' => $lastName,
                 'email' => strtolower($firstName).'@spatie.be',
@@ -32,11 +32,11 @@ class BackUserSeeder extends DatabaseSeeder
         });
     }
 
-    public function createBackUser(array $attributes = []): User
+    public function createAdministrator(array $attributes = []): Administrator
     {
         $person = faker()->person();
 
-        return User::create($attributes + [
+        return Administrator::create($attributes + [
             'first_name' => $person['firstName'],
             'last_name' => $person['lastName'],
             'email' => $person['email'],
